requirements.txt
geographiclib==2.0
numpy==1.24.2
paho_mqtt==1.6.1
